Move the object list reader out of the models to make it cleaner
